requirements.txt
pandas
numpy
openpyxl
supabase
python-dotenv